LPS-193319 Move WorkflowDefinitionManager to workflow-api module.
