+++ modules/apps/message-boards/message-boards-moderation/src/main/java/com/liferay/message/boards/moderation/internal/configuration/persistence/listener/MBModerationGroupConfigurationModelListener.java
@@ -17,7 +17,7 @@ import com.liferay.portal.kernel.service.WorkflowDefinitionLinkLocalService;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.HashMapDictionary;
 import com.liferay.portal.kernel.workflow.WorkflowDefinition;
-import com.liferay.portal.kernel.workflow.WorkflowDefinitionManager;
+import com.liferay.portal.workflow.manager.WorkflowDefinitionManager;
 
 import java.util.Dictionary;
