+++ modules/apps/message-boards/message-boards-moderation/src/main/java/com/liferay/message/boards/moderation/internal/upgrade/registry/MBModerationUpgradeStepRegistrator.java
@@ -7,9 +7,9 @@ package com.liferay.message.boards.moderation.internal.upgrade.registry;
 
 import com.liferay.message.boards.moderation.internal.upgrade.v1_0_0.MBModerationWorkflowDefinitionUpgradeProcess;
 import com.liferay.portal.kernel.service.CompanyLocalService;
-import com.liferay.portal.kernel.workflow.WorkflowDefinitionManager;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
 import com.liferay.portal.workflow.kaleo.runtime.WorkflowEngine;
+import com.liferay.portal.workflow.manager.WorkflowDefinitionManager;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
