+++ modules/apps/message-boards/message-boards-moderation/src/main/java/com/liferay/message/boards/moderation/internal/upgrade/v1_0_0/MBModerationWorkflowDefinitionUpgradeProcess.java
@@ -12,8 +12,8 @@ import com.liferay.portal.kernel.service.CompanyLocalService;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.workflow.WorkflowDefinition;
-import com.liferay.portal.kernel.workflow.WorkflowDefinitionManager;
 import com.liferay.portal.kernel.workflow.WorkflowException;
+import com.liferay.portal.workflow.manager.WorkflowDefinitionManager;
 
 /**
  * @author Eduardo García
